moves availability.js into a separate pack and refactors it
