+++ app/javascript/availability/item_types.json
