+++ app/javascript/availability/libraries_locations.json
