+++ app/javascript/packs/application.js
@@ -7,10 +7,9 @@
 // To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
 // layout file, like app/views/layouts/application.html.erb
 
-
-
 import bookCovers from "../book_covers";
 import askALibrarian from "../ask_a_librarian";
+import availability from "../availability";
 require.context('../psulib_blacklight/images/', true);
 
 import Rails from 'rails-ujs';
@@ -26,6 +25,7 @@ import 'blacklight_overrides'
 
 document.addEventListener("turbolinks:load", function() {
     askALibrarian.start();
+    availability.executeAvailability();
     bookCovers.start();
     Blacklight.doBookmarkToggleBehavior();
 });
