+++ app/javascript/psulib_blacklight/javascripts/index.js
@@ -1,5 +1,4 @@
 import './show_default';
-import './availability';
 
 // Otherwise bootstrap-select won't fire on turbolinked clicks to Advance Search
 $(document).on('turbolinks:load', function() {
